requirements.txt
opencv-python
h5py
numpy
pillow